# gpt-api/openapi.yaml
openapi: 3.1.0
info:
  title: Pegasus GPT API
  version: 1.0.0
  description: Full-stack system management API for Pegasus Wingbot
servers:
  - url: https://gpt.1tsjeremiah.com
paths:
...
  /gpt/chat:
    post:
      operationId: chatGpt
      summary: Send a message to OpenAI GPT model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  example: "Summarize this week's server activity."
              required: [prompt]
      responses:
        '200':
          description: GPT response
          content:
            text/plain:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
security:
  - bearerAuth: []